# Erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/talde2erronka2back/Erronka2/public/api/fichas/editatu": {
            "put": {
                "tags": [
                    "Fitxak"
                ],
                "summary": "Fitxak editatu",
                "operationId": "866caf7f80ea1ce95fdfbfe3ed72f523",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "izena",
                                    "abizena",
                                    "telefonoa",
                                    "azal_sentikorra"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "izena": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "abizena": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "telefonoa": {
                                        "type": "string",
                                        "example": "555123456"
                                    },
                                    "azal_sentikorra": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fitxa editatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/fichas/txertatu": {
            "post": {
                "tags": [
                    "Fitxak"
                ],
                "summary": "Fitxak txertatu",
                "operationId": "55be0dc95abbc597449e2530f8ca57ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "izena",
                                    "abizena",
                                    "telefonoa",
                                    "azal_sentikorra"
                                ],
                                "properties": {
                                    "izena": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "abizena": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "telefonoa": {
                                        "type": "string",
                                        "example": "555123456"
                                    },
                                    "azal_sentikorra": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fitxa txertatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/fichas/ezabatu": {
            "put": {
                "tags": [
                    "Fitxak"
                ],
                "summary": "Fitxak ezabatu",
                "operationId": "4f85d7331a7cdfbe3f437817464b5e03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fitxa ezabatu da."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}